fix(interface): :bug: properly display active state on restore
if the intended behaviour was for the hide radio to be restored to active, the only thing that would actually happen is that checked = false would be applied to the show radio.

the hide radio would never be set to checked, and hence the active state (ie bolding) would never apply.