fix(safari): :alien: add internal integration token fallback
browser.identity is not supported in Safari, which breaks the current OAuth2 implementation.

from what I can tell, there is also no way to produce a constant url to act as the redirect_uri, which makes it impossible to implementent a 'secure' OAuth2 PKCE flow.

as a temporary stop-gap, the extension will fall back to a text input for an internal integration token if browser.identity is unavailable.